requirements.txt
python-dotenv==1.1.1
python-telegram-bot==22.5
